fix navigation buttons in group questionnaires
