+++ NinchatSDKSwift/Implementations/View/Coordinator/Coordinator.swift
@@ -67,7 +67,7 @@ final class NINCoordinator: Coordinator {
     internal lazy var questionnaireViewController: NINQuestionnaireViewController = {
         let questionnaireViewController: NINQuestionnaireViewController = storyboard.instantiateViewController()
         questionnaireViewController.session = session
-        questionnaireViewController.pageNumber = 0
+        questionnaireViewController.pageNumber = 10
         questionnaireViewController.finishQuestionnaire = {
             DispatchQueue.main.async {
                 #warning("Finish the questionnaire")
