+++ NinchatSDKSwift/Implementations/View/ViewController/NINQuestionnaireViewController.swift
@@ -8,15 +8,25 @@ import UIKit
 
 final class NINQuestionnaireViewController: UIViewController, ViewController {
 
-    private var elements: [QuestionnaireElement] {
+    private var _configuration: QuestionnaireConfiguration {
+        if let audienceQuestionnaire = session.sessionManager.siteConfiguration.preAudienceQuestionnaire {
+            guard audienceQuestionnaire.count > self.pageNumber else { fatalError("Invalid number of questionnaires configurations") }
+
+            return audienceQuestionnaire[self.pageNumber]
+        }
+        fatalError("Configuration for the page number: \(self.pageNumber ?? nil) is not exits")
+    }
+    private var _elements: [QuestionnaireElement] {
         if let audienceQuestionnaire = session.sessionManager.siteConfiguration.preAudienceQuestionnaire {
             let views = QuestionnaireElementConverter(configurations: audienceQuestionnaire).elements
-            guard views.count > self.pageNumber else { fatalError("Invalid number of questionnaires") }
+            guard views.count > self.pageNumber else { fatalError("Invalid number of questionnaires views") }
 
             return views[self.pageNumber]
         }
         return []
     }
+    private var configuration: QuestionnaireConfiguration!
+    private var elements: [QuestionnaireElement] = []
 
     // MARK: - ViewController
 
@@ -43,6 +53,8 @@ final class NINQuestionnaireViewController: UIViewController, ViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
+        elements = _elements
+        configuration = _configuration
         contentView = generateTableView(isHidden: false)
     }
 
@@ -57,6 +69,8 @@ final class NINQuestionnaireViewController: UIViewController, ViewController {
     }
 
     private func updateLayout() {
+        elements = _elements
+        configuration = _configuration
         contentView?.hide(true, andCompletion: { [weak self] in
             DispatchQueue.main.async {
                 self?.contentView?.removeFromSuperview()
@@ -71,6 +85,7 @@ extension NINQuestionnaireViewController {
     private func generateTableView(isHidden: Bool) -> UITableView {
         let view = UITableView(frame: .zero)
         view.register(QuestionnaireCell.self)
+        view.registerClass(QuestionnaireNavigationCell.self)
 
         view.separatorStyle = .none
         view.allowsSelection = false
@@ -84,25 +99,36 @@ extension NINQuestionnaireViewController {
 
 extension NINQuestionnaireViewController: UITableViewDataSource, UITableViewDelegate {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        self.elements[indexPath.row].height
+        if indexPath.row == self.elements.count {
+            return 65.0
+        }
+        return self.elements[indexPath.row].height
     }
 
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        elements.count
+        self.elements.count + 1
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        if indexPath.row == self.elements.count {
+            /// Show navigation buttons
+            let cell: QuestionnaireNavigationCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+            cell.configuration = self.configuration
+            cell.onNextButtonTapped = { [weak self] questionnaire in
+                self?.pageNumber += 1
+                self?.updateLayout()
+            }
+            cell.onBackButtonTapped = { [weak self] questionnaire in
+                self?.pageNumber -= 1
+                self?.updateLayout()
+            }
+            return cell
+        }
+
+        /// Show questionnaire items
         let cell: QuestionnaireCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
         let view = self.elements[indexPath.row]
         view.overrideAssets(with: self.session, isPrimary: false)
-        (view as? QuestionnaireElementWithNavigationButtons)?.onNextButtonTapped = { [weak self] questionnaire in
-            self?.pageNumber += 1
-            self?.updateLayout()
-        }
-        (view as? QuestionnaireElementWithNavigationButtons)?.onBackButtonTapped = { [weak self] questionnaire in
-            self?.pageNumber -= 1
-            self?.updateLayout()
-        }
         self.layoutSubview(view, parent: cell.content)
 
         return cell
